3/13/16 Update
The rotating bug still hasn’t been fixed, nor has the count down for
enemy spawns…

But now there is a placeholder game over screen, and spawning occurs
with a pause of 3 to 6 seconds between enemies on each row. This is far
from perfect (there’s still a risk of spawns on all three rows
happening so close together that there’s no fair way for the player to
proceed), but WAY better than before.

Sometimes the top row completely disappears for me. I thought that
maybe my destroyer block was killing the top spawner and its children,
but OnDestroy didn’t seem to trigger. I’ll take a closer look at it
tomorrow, but feel free to look at that as well.